extracted classes from storage.py and first minimal Redis sync for app.storage.general
