+++ nicegui/ui_run_with.py
@@ -4,10 +4,11 @@ from typing import Literal, Optional, Union
 
 from fastapi import FastAPI
 
-from . import core, storage
+from . import core
 from .air import Air
 from .language import Language
 from .nicegui import _shutdown, _startup
+from .persistence import storage
 
 
 def run_with(
